refact: move SerializableList-related structures to list package
