+++ loadable/src/main/java/com/jeanbarrossilva/loadable/list/SerializableList.kt
@@ -1,4 +1,4 @@
-package com.jeanbarrossilva.loadable.type
+package com.jeanbarrossilva.loadable.list
 
 import java.io.Serializable
